requirements.txt
requests
pandas
sqlalchemy
scikit-learn
xgboost
numpy
python-dotenv
schedule
pytz
matplotlib
streamlit
